Reconstruct and add more
Use this.whiteKeyWidth and this.blackKeyWidth to store the value
of the width of keys,add a getKeyLeftPosition method, which
is called from only a few places,and add more code for drawing
note. Now a note is generated when key is pressed down, but it
neither move or disappear,which will be implemented later.