Add a new feature
Now the notes go up in normal mode and recording mode, but go down
in replaying mode. I add method for Plotter that is used for
Keyboard to inform plotter the start and end of replaying, and do
some modification in method update and createNote, and I try add
notes which is in the horizon from the beginning of the replay,and
do something in Keyboard.endRecord to ensure that the Sequences
returned always have all key released.For some issues mentioned
above, I changed the value of Sequences.actions.key from key(string)
to key_id(int), which is more intuitive, because in _replay, the
value is used to call _keyDown and _keyUp but not keyDown and keyUp.