Fix a bug
A small bug, but a lot of time :( .
1. I add an empty Sequence check before replay it, which cause an
error.
2. split the process of note actions which is
Plotter.noteDisplayTime long time ahead of key up down and note
play, I worry about the synchronization between the two threads.
3. I move code changing key_playing to keyDown and keyUp because
the it is used to avoid continuous keyDown calling when the key is
pressed down and hold by user.
4. move code to create or modify notes to two stand-alone methods,
in order to call them in _replay.