Add a new feature
When practising on the keyboard, I find it is inconvenient to push
down a black key, right after its conjugate white key is release,
so I do some modification. Now when the shift is pressed, all
pressed key is changed to their correspond keys. Whether there is
bugs is not tested.(and a return true is added at the end of keyUp.